+++ src/pyledctrl/config.py
@@ -1,3 +1,0 @@
-"""Configuration constants for PyLedCtrl."""
-
-DEFAULT_BAUD = 115200