+++ src/pyledctrl/utils.py
@@ -1,29 +1,40 @@
 """Utility functions for PyLedCtrl."""
 
-import glob
 import os
-import shutil
 import sys
-import tempfile
 
 from itertools import tee
+from typing import cast, Callable, Iterable, Tuple, TypeVar, overload
 
-from pyledctrl.config import DEFAULT_BAUD
+
+T = TypeVar("T")
+T2 = TypeVar("T2")
+Tup = TypeVar("Tup", bound="Tuple")
 
 
-def consecutive_pairs(iterable):
+def consecutive_pairs(it: Iterable[T]) -> Iterable[Tuple[T, T]]:
     """Given an iterable, returns a generator that generates consecutive
     pairs of items from the iterable.
 
     Parameters:
-        iteable (Iterable): the iterable
+        it: the iterable
 
     Yields:
-        (object, object): pairs of consecutive items from the iterable
+        pairs of consecutive items from the iterable
     """
-    a, b = tee(iterable)
+    a, b = tee(it)
     next(b, None)
     return zip(a, b)
+
+
+@overload
+def ensure_tuple(obj: Tup) -> Tup:
+    ...
+
+
+@overload
+def ensure_tuple(obj: T) -> Tuple[T]:
+    ...
 
 
 def ensure_tuple(obj):
@@ -33,36 +44,39 @@ def ensure_tuple(obj):
     return (obj if isinstance(obj, tuple) else obj,)
 
 
-def error(message, fatal=False):
+def error(message: str, fatal: bool = False) -> None:
     """Prints an error message to stderr.
 
-    :param message: The message to print
-    :param fatal: Whether to terminate the script after the error message
+    Parameters:
+        message: The message to print
+        fatal: Whether to terminate the script after the error message
     """
     print(message, file=sys.stderr)
     if fatal:
         sys.exit(1)
 
 
-def first(iterable):
-    """Returns the first element from the given iterable. Raises ``ValueError``
-    if the iterable is empty.
+def first(iterable: Iterable[T]) -> T:
+    """Returns the first element from the given iterable.
+
+    Raises:
+        ValueError: if the iterable is empty.
     """
     for item in iterable:
         return item
     raise ValueError("iterable is empty")
 
 
-def format_frame_count(frames, fps):
+def format_frame_count(frames: int, *, fps: int) -> str:
     """Formats a time instant given as the number of frames since T=0 into
     a conventional minutes:seconds+frames representation.
 
     Parameters:
-        frames (int): the number of frames
-        fps (int): the number of frames per second
+        frames: the number of frames
+        fps: the number of frames per second
 
     Returns:
-        str: the formatted representation of the frame count
+        the formatted representation of the frame count
     """
     seconds, residual = divmod(frames, fps)
     minutes, seconds = divmod(seconds, 60)
@@ -71,50 +85,14 @@ def format_frame_count(frames, fps):
     )
 
 
-def get_serial_port_filename(port=None):
-    """Returns the serial port filename from the given string, handling
-    defaults gracefully.
-
-    :param port: the serial port, ``None`` means to use the first USB serial
-        port
-    """
-    if port is None:
-        ttyusb = glob.glob("/dev/ttyUSB*") + ["/dev/ttyUSB0"]
-        return ttyusb[0]
-    else:
-        return str(port)
-
-
-def get_serial_connection(port, baud=None):
-    """Returns a serial connection object for the given port and baud rate,
-    handling defaults gracefully.
-
-    :param port: the serial port, ``None`` means to use the first USB serial
-        port
-    :param baud: the baud rate, ``None`` uses the default baud rate
-    """
-    # Lazy import. This is important since groundctrl is not a strict
-    # dependency of pyledctrl
-    from groundctrl.serial_port import SerialPort
-
-    return SerialPort(port=get_serial_port_filename(port), baud=baud or DEFAULT_BAUD)
-
-
-def iterbytes(fp):
-    """Iterates over the bytes of a file-like object."""
-    while True:
-        b = fp.read(1)
-        if not b:
-            return
-        yield b
-
-
 _last_default = object()
 
 
-def last(iterable):
-    """Returns the last element from the given iterable. Raises ``ValueError``
-    if the iterable is empty.
+def last(iterable: Iterable[T]) -> T:
+    """Returns the last element from the given iterable.
+
+    Raises:
+        ValueError: if the iterable is empty
     """
     last = _last_default
     for last in iterable:
@@ -122,10 +100,10 @@ def last(iterable):
     if last is _last_default:
         raise ValueError("iterable is empty")
     else:
-        return last
+        return cast(T, last)
 
 
-def memoize(func):
+def memoize(func: Callable[[T], T2]) -> Callable[[T], T2]:
     """Single-argument memoization decorator for a function. Caches the results
     of the function in a dictionary.
     """
@@ -133,14 +111,14 @@ def memoize(func):
     class memodict(dict):
         __slots__ = ()
 
-        def __missing__(self, key):
+        def __missing__(self, key: T):
             self[key] = ret = func(key)
             return ret
 
     return memodict().__getitem__
 
 
-def parse_as_frame_count(value, fps):
+def parse_as_frame_count(value: str, *, fps: int) -> int:
     """Parses the given input string containing the representation of a time
     instant in the usual ``minutes:seconds+frames`` format into an absolute
     frame count.
@@ -152,11 +130,11 @@ def parse_as_frame_count(value, fps):
     assumed to be zero.
 
     Parameters:
-        value (str): the input string to parse
-        fps (int): the number of frames per second
+        value: the input string to parse
+        fps: the number of frames per second
 
     Returns:
-        int: the absolute frame count parsed out from the string
+        the absolute frame count parsed out from the string
     """
     minutes, _, seconds = value.rpartition(":")
     minutes = float(minutes) if minutes else 0
@@ -178,45 +156,3 @@ def replace_extension(filename: str, ext: str) -> str:
     """
     base, _ = os.path.splitext(filename)
     return base + ext
-
-
-class _TemporaryDirectory:
-    """Create and return a temporary directory.  This has the same
-    behaviour as mkdtemp but can be used as a context manager.
-    Backported from Python 3.x. For example:
-
-        with TemporaryDirectory() as tmpdir:
-            ...
-
-    Upon exiting the context, the directory and everything contained
-    in it are removed.
-    """
-
-    # Handle mkdtemp raising an exception
-    name = None
-    _closed = False
-
-    def __init__(self, suffix="", prefix=tempfile.template, dir=None, keep=False):
-        self.name = tempfile.mkdtemp(suffix, prefix, dir)
-        self.keep = keep
-
-    def __repr__(self):
-        return "<{} {!r}>".format(self.__class__.__name__, self.name)
-
-    def __enter__(self):
-        return self.name
-
-    def __exit__(self, exc, value, tb):
-        self.cleanup()
-
-    def cleanup(self):
-        if self.name is not None and not self._closed:
-            if not self.keep:
-                shutil.rmtree(self.name)
-            self._closed = True
-
-
-try:
-    from tempfile import TemporaryDirectory
-except ImportError:
-    TemporaryDirectory = _TemporaryDirectory